+++ include/thermalForce.h
@@ -75,17 +75,17 @@ struct thermalForce {
       // std:cout << " grad Ti interp " << std::endl;
       interp2dVector(&gradTi[0], p->xprevious[indx], p->yprevious[indx], p->zprevious[indx], nR_gradT, nZ_gradT,
                      gradTGridr, gradTGridz, gradTiR, gradTiZ, gradTiT);
-      //std::cout << "Position r z" << sqrt(p->xprevious*p->xprevious + p->yprevious*p->yprevious) << " " << p->zprevious << std::endl;
-      //std::cout << "grad Ti " << std::copysign(1.0,gradTi[0])*sqrt(gradTi[0]*gradTi[0] + gradTi[1]*gradTi[1]) << " " << gradTi[2] << std::endl;
+      //std::cout << "Position r z" << std::sqrt(p->xprevious*p->xprevious + p->yprevious*p->yprevious) << " " << p->zprevious << std::endl;
+      //std::cout << "grad Ti " << std::copysign(1.0,gradTi[0])*std::sqrt(gradTi[0]*gradTi[0] + gradTi[1]*gradTi[1]) << " " << gradTi[2] << std::endl;
       interp2dVector(&gradTe[0], p->xprevious[indx], p->yprevious[indx], p->zprevious[indx], nR_gradT, nZ_gradT,
                      gradTGridr, gradTGridz, gradTeR, gradTeZ, gradTeT);
       mu = p->amu[indx] / (background_amu + p->amu[indx]);
       alpha = p->charge[indx] * p->charge[indx] * 0.71;
-      beta = 3 * (mu + 5 * sqrtf(2.0) * p->charge[indx] * p->charge[indx] * (1.1 * powf(mu, (5 / 2)) - 0.35 * powf(mu, (3 / 2))) - 1) / (2.6 - 2 * mu + 5.4 * powf(mu, 2));
+      beta = 3 * (mu + 5 * std::sqrt(2.0) * p->charge[indx] * p->charge[indx] * (1.1 * std::pow(mu, (5 / 2)) - 0.35 * std::pow(mu, (3 / 2))) - 1) / (2.6 - 2 * mu + 5.4 * std::pow(mu, 2));
 
       interp2dVector(&B[0], p->xprevious[indx], p->yprevious[indx], p->zprevious[indx], nR_Bfield, nZ_Bfield,
                      BfieldGridRDevicePointer, BfieldGridZDevicePointer, BfieldRDevicePointer, BfieldZDevicePointer, BfieldTDevicePointer);
-      Bmag = sqrtf(B[0] * B[0] + B[1] * B[1] + B[2] * B[2]);
+      Bmag = std::sqrt(B[0] * B[0] + B[1] * B[1] + B[2] * B[2]);
       B_unit[0] = B[0] / Bmag;
       B_unit[1] = B[1] / Bmag;
       B_unit[2] = B[2] / Bmag;
@@ -108,25 +108,25 @@ struct thermalForce {
       //std::cout << "ETG " << dv_ETG[0] << " " << dv_ETG[1] << " " << dv_ETG[2] << std::endl;
       //std::cout << "v before thermal force " << p->vx[indx] << " " << p->vy[indx] << " " << p->vz[indx] << std::endl;
       /*
-    float theta = atan2(p->yprevious,p->xprevious);
+    float theta = std::atan2(p->yprevious,p->xprevious);
     float Ar = -1;
     float At = 0.0;
     float Az = 1;
-    gradTi[0] = cos(theta)*Ar - sin(theta)*At;
-    gradTi[1] = sin(theta)*Ar + cos(theta)*At;
+    gradTi[0] = std::cos(theta)*Ar - std::sin(theta)*At;
+    gradTi[1] = std::sin(theta)*Ar + std::cos(theta)*At;
     gradTi[2] = Az;
     */
       float vx = p->vx[indx];
       float vy = p->vy[indx];
       float vz = p->vz[indx];
-      vNorm = sqrt(vx * vx + vy * vy + vz * vz);
+      vNorm = std::sqrt(vx * vx + vy * vy + vz * vz);
       p->vD[indx] = dv_ITG[2];
       //std::cout << "gradTi Parallel " << gradTiPar << std::endl;
       //std::cout << "gradTi Parallel " << gradTi[0]<<gradTi[1]<<gradTi[2] << std::endl;
       //p->vx[indx] = p->vx[indx] +dv_ITG[0];//alpha*(gradTe[0])
       //p->vy[indx] = p->vy[indx] +dv_ITG[1];//alpha*(gradTe[1])
       //p->vz[indx] = p->vz[indx] +dv_ITG[2];//alpha*(gradTe[2])
-      //vNorm2 = sqrt(p->vx[indx]*p->vx[indx] + p->vy[indx]*p->vy[indx] + p->vz[indx]*p->vz[indx]);
+      //vNorm2 = std::sqrt(p->vx[indx]*p->vx[indx] + p->vy[indx]*p->vy[indx] + p->vz[indx]*p->vz[indx]);
       //SFT
       float k1 = dv_ITG[2] - dt * p->nu_s[indx] * (dv_ITG[2]);
       p->vx[indx] = vx + dv_ITG[0]; ///velocityCollisionsNorm;
