+++ src/interp2d.cpp
@@ -75,8 +75,8 @@ float interp2dCombined(float x, float y, float z, int nx, int nz,
 #endif
     float d_dim1 = gridx[1] - gridx[0];
     float dz = gridz[1] - gridz[0];
-    int i = floor((dim1 - gridx[0]) / d_dim1); //addition of 0.5 finds nearest gridpoint
-    int j = floor((z - gridz[0]) / dz);
+    int i = std::floor((dim1 - gridx[0]) / d_dim1); //addition of 0.5 finds nearest gridpoint
+    int j = std::floor((z - gridz[0]) / dz);
 
     //float interp_value = data[i + j*nx];
     if (i < 0)
@@ -117,9 +117,9 @@ float interp3d(float x, float y, float z, int nx, int ny, int nz,
   float dy = gridy[1] - gridy[0];
   float dz = gridz[1] - gridz[0];
 
-  int i = floor((x - gridx[0]) / dx); //addition of 0.5 finds nearest gridpoint
-  int j = floor((y - gridy[0]) / dy);
-  int k = floor((z - gridz[0]) / dz);
+  int i = std::floor((x - gridx[0]) / dx); //addition of 0.5 finds nearest gridpoint
+  int j = std::floor((y - gridy[0]) / dy);
+  int k = std::floor((z - gridz[0]) / dz);
   //std::cout << "dxyz ijk " << dx << " "<<dy << " " << dz<< " " << i
   //  << " " << j << " " << k << std::endl;
   if (i < 0)
@@ -184,7 +184,7 @@ void interp2dVector(float *field, float x, float y, float z, int nx, int nz,
   float At = interp2dCombined(x, y, z, nx, nz, gridx, gridz, datat);
   field[2] = interp2dCombined(x, y, z, nx, nz, gridx, gridz, dataz);
 #if USECYLSYMM > 0
-  float theta = atan2f(y, x);
+  float theta = std::atan2(y, x);
   field[0] = std::cos(theta) * Ar - std::sin(theta) * At;
   field[1] = std::sin(theta) * Ar + std::cos(theta) * At;
 #else
